Changed heightmap from int values to byte values.

+++ CNCMaps.Engine/Generator/GeneratorEngine.cs
@@ -32,7 +32,7 @@ namespace CNCMaps.Engine.Generator {
 		internal ushort Height;
 		internal ushort Width;
 		internal PerlinNoise Noise { get; }
-		internal int[,] HeightLayout { get; set; }
+		internal byte[,] HeightLayout { get; set; }
 		private Logger _logger { get; }
 
 		public virtual bool GenerateMap() {
@@ -107,12 +107,12 @@ namespace CNCMaps.Engine.Generator {
 
 		internal void GenerateHeightLayout() {
 			_logger.Debug("Generating height layout");
-			HeightLayout = new int[TileLayer.Height, TileLayer.Width];
+			HeightLayout = new byte[TileLayer.Height, TileLayer.Width];
 			var nv = 0d;
 			for (int y = 0; y < TileLayer.Height; y++) {
 				for (int x = 0; x < TileLayer.Width; x++) {
 					nv = Noise.Noise(x * NoiseOffset, y * NoiseOffset, 0d) + 1d;
-					HeightLayout[y,x] = (int)(nv * 128);
+					HeightLayout[y,x] = (byte)(nv * 128);
 				}
 			}
 			DebugLayoutHeight();
